refactor get assessments along dimensions for resource for modularity
